Move RWLock into its own module, mutually independent GRPC2 module and GRPC module.
